app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "The backend API for the Next-Gen Support Agent"
authors = [
    {name = "Yunus Emre Gültepe", email = "yemregultepe@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "langchain>=0.3.27,<0.4.0",
    "langchain-community>=0.3.27,<0.4.0",
    "pypdf>=5.9.0,<6.0.0",
    "sentence-transformers>=5.0.0,<6.0.0",
    "chromadb-client>=1.0.15,<2.0.0",
    "langchain-huggingface>=0.3.1,<0.4.0",
    "flashrank>=0.2.10,<0.3.0",
    "langchain-chroma>=0.2.5,<0.3.0",
    "langchain-openai>=0.3.28,<0.4.0",
    "langsmith>=0.4.12,<0.5.0",
    "langchain-redis>=0.2.3,<0.3.0",
]

[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Yunus Emre Gültepe <yemregultepe@gmail.com>"]
# --- THIS IS THE FIX ---
# Explicitly tell Poetry where the source code for the "app" package is.
packages = [{include = "app", from = ".."}]

[tool.poetry.group.dev.dependencies]
types-redis = "^4.6.0.20240417"
types-requests = "^2.31.0.20240406"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
